default selection when user comes from other page error messages when group assignment is not successful

+++ workspace/enwida/src/main/java/de/enwida/web/dao/implementation/UserDao.java
@@ -32,43 +32,6 @@ public class UserDao extends BaseDao<User> implements IUserDao {
 	@Autowired
 	private DriverManagerDataSource datasource;
 	
-	public List<User> findAllUsersWithPermissions(){
-		ArrayList<User> users = new ArrayList<User>();
-		String sql = "SELECT * FROM users";
-		 
-		Connection conn = null;
- 
-		try {
-			conn = datasource.getConnection();
-			PreparedStatement ps = conn.prepareStatement(sql);
-			User user = null;
-			ResultSet rs = ps.executeQuery();
-			while (rs.next()) {
-				user = new User(
-					rs.getLong("user_id"),
-					rs.getString("user_name"), 
-					rs.getString("user_password"), 
-					rs.getBoolean("enabled")
-				);
-				;
-				user.setFirstName(rs.getString("first_name"));
-				user.setLastName(rs.getString("last_name"));
-				users.add(loadUserFromDB(user));
-			}
-			rs.close();
-			ps.close();
-		} catch (SQLException e) {
-			throw new RuntimeException(e);
-		} finally {
-			if (conn != null) {
-				try {
-				conn.close();
-				} catch (SQLException e) {}
-			}
-		}
-
-		return users;
-	}
 	
 	public List<User> findAllUsers(){
 		ArrayList<User> users = new ArrayList<User>();
@@ -89,6 +52,12 @@ public class UserDao extends BaseDao<User> implements IUserDao {
 					rs.getBoolean("enabled")
 				);
 				;
+                user.setFirstName(rs.getString("first_name"));
+                user.setLastName(rs.getString("last_name"));
+                ArrayList<Group> groups = getUserGroups(user.getUserID());
+                user.setGroups(groups);
+                ArrayList<Role> roles = getUserRoles(user.getUserID());
+                user.setRoles(roles);
 				users.add(user);
 			}
 			rs.close();
@@ -106,7 +75,39 @@ public class UserDao extends BaseDao<User> implements IUserDao {
 		return users;
 	}
 	
-	public boolean createUser(User user){
+	private ArrayList<Role> getUserRoles(Long userID) {
+	    String sql = "select DISTINCT ON (role_id) roles.role_id,roles.role_name FROM roles " +
+	    		"INNER JOIN group_role ON group_role.role_id=roles.role_id " +
+	    		"INNER JOIN user_group ON user_group.group_id=group_role.group_id " +
+	    		" where user_group.user_id=?";
+        Connection conn = null;
+        ArrayList<Role> roles = new ArrayList<Role>();
+        try {
+            conn = datasource.getConnection();
+            PreparedStatement ps = conn.prepareStatement(sql);
+            ps.setLong(1, userID);
+            ResultSet rs = ps.executeQuery();
+            while (rs.next()) {
+                Role role = new Role();
+                role.setRoleID(rs.getLong("role_id"));
+                role.setRoleName(rs.getString("role_name"));
+                roles.add(role);
+            }
+            rs.close();
+            ps.close();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        } finally {
+            if (conn != null) {
+                try {
+                conn.close();
+                } catch (SQLException e) {}
+            }
+        }
+        return roles;
+    }
+
+    public boolean createUser(User user){
 		
 		String sql = "INSERT INTO users(user_name, user_password, first_name, last_name, enabled, joining_date) VALUES (?, ?, ?, ?, ?, current_date)";
 		 
@@ -135,44 +136,6 @@ public class UserDao extends BaseDao<User> implements IUserDao {
 		return true;
 	}
 
-	public User loadUserFromDB(User user) {
-		
-		String sql = "SELECT users.user_id,user_name,user_password,enabled,string_agg(roles.role_name, ', ')as permissions" +
-				" FROM users  INNER JOIN user_roles ON users.user_id=user_roles.user_id" +
-				" INNER JOIN roles ON roles.role_id=user_roles.role_id " +
-				" where users.user_name=? group by users.user_id,user_name,user_password,enabled";
-
-		Connection conn = null;
- 
-		try {
-			conn = datasource.getConnection();
-			PreparedStatement ps = conn.prepareStatement(sql);
-			ps.setString(1, user.getUserName());
-			ResultSet rs = ps.executeQuery();
-			while (rs.next()) {
-				String[] authority= rs.getString("permissions").split(",");
-				for (String auth : authority) {
-					UserPermission permission=new UserPermission(auth);	
-					user.getUserPermissionCollection().add(permission);
-				}
-				user.setUserID(rs.getLong("user_id"));
-				user.setUserName(rs.getString("user_name"));
-				user.setPassword(rs.getString("user_password"));
-				user.setEnabled(rs.getBoolean("enabled"));
-			}
-			rs.close();
-			ps.close();
-		} catch (SQLException e) {
-			throw new RuntimeException(e);
-		} finally {
-			if (conn != null) {
-				try {
-				conn.close();
-				} catch (SQLException e) {}
-			}
-		}
-		return user;
-	}
 	
 	public void addPermission(int userID, int roleID) {
 		
@@ -393,6 +356,10 @@ public class UserDao extends BaseDao<User> implements IUserDao {
                 user.setLastName(rs.getString("last_name"));
                 user.setJoinDate(rs.getString("joining_date"));
                 user.setTel(rs.getString("tel"));
+                ArrayList<Group> groups = getUserGroups(user.getUserID());
+                user.setGroups(groups);
+                ArrayList<Role> roles = getUserRoles(user.getUserID());
+                user.setRoles(roles);
 			}
 			rs.close();
 			ps.close();
@@ -693,6 +660,10 @@ public class UserDao extends BaseDao<User> implements IUserDao {
                 user.setLastName(rs.getString("last_name"));
                 user.setJoinDate(rs.getString("joining_date"));
                 user.setTel(rs.getString("tel"));
+                ArrayList<Group> groups = getUserGroups(user.getUserID());
+                user.setGroups(groups);
+                ArrayList<Role> roles = getUserRoles(user.getUserID());
+                user.setRoles(roles);
             }
             rs.close();
             ps.close();
@@ -709,9 +680,9 @@ public class UserDao extends BaseDao<User> implements IUserDao {
         return user;
     }
 
-    public boolean assignUserToGroup(int userID, int groupID) {
-        if(userID==0){
-            return false;
+    public String assignUserToGroup(int userID, int groupID) {
+        if(userID==0 || groupID==0){
+            return "Invalid userID  or groupID ";
         }
         
         
@@ -727,7 +698,7 @@ public class UserDao extends BaseDao<User> implements IUserDao {
             ps.executeUpdate();
             ps.close();
         } catch (SQLException e) {
-            return false;
+            return e.getLocalizedMessage();
         } finally {
             if (conn != null) {
                 try {
@@ -735,12 +706,12 @@ public class UserDao extends BaseDao<User> implements IUserDao {
                 } catch (SQLException e) {}
             }
         }
-        return true;
+        return "OK";
     }
 
-    public boolean deassignUserToGroup(int userID, int groupID) {
-        if(userID==0){
-            return false;
+    public String deassignUserToGroup(int userID, int groupID) {
+        if(userID==0 || groupID==0){
+            return "Invalid userID  or groupID ";
         }
         
         String sql = "DELETE FROM user_group WHERE (user_id=? and group_id=?)";
@@ -754,7 +725,7 @@ public class UserDao extends BaseDao<User> implements IUserDao {
             ps.executeUpdate();
             ps.close();
         } catch (SQLException e) {
-            return false;
+            return e.getLocalizedMessage();
         } finally {
             if (conn != null) {
                 try {
@@ -762,12 +733,12 @@ public class UserDao extends BaseDao<User> implements IUserDao {
                 } catch (SQLException e) {}
             }
         }       
-        return true;
+        return "OK";
     }
 
-    public boolean assignRoleToGroup(int roleID, int groupID) {
-        if(roleID==0){
-            return false;
+    public String assignRoleToGroup(int roleID, int groupID) {
+        if(roleID==0 || groupID==0){
+            return "Invalid roleID  or groupID ";
         }
         
         
@@ -783,7 +754,7 @@ public class UserDao extends BaseDao<User> implements IUserDao {
             ps.executeUpdate();
             ps.close();
         } catch (SQLException e) {
-            return false;
+            return e.getLocalizedMessage();
         } finally {
             if (conn != null) {
                 try {
@@ -791,12 +762,12 @@ public class UserDao extends BaseDao<User> implements IUserDao {
                 } catch (SQLException e) {}
             }
         }
-        return true; 
+        return "OK"; 
     }
 
-    public boolean deassignRoleToGroup(int roleID, int groupID) {
-        if(roleID==0){
-            return false;
+    public String deassignRoleToGroup(int roleID, int groupID) {
+        if(roleID==0 || groupID==0){
+            return "Invalid roleID  or groupID ";
         }
         
         String sql = "DELETE FROM group_roles WHERE (group_id=? and role_id=?)";
@@ -810,7 +781,7 @@ public class UserDao extends BaseDao<User> implements IUserDao {
             ps.executeUpdate();
             ps.close();
         } catch (SQLException e) {
-            return false;
+            return e.getLocalizedMessage();
         } finally {
             if (conn != null) {
                 try {
@@ -818,7 +789,7 @@ public class UserDao extends BaseDao<User> implements IUserDao {
                 } catch (SQLException e) {}
             }
         }       
-        return true;
+        return "OK";
     }
 
     public List<Role> getAllRolesWithGroups() {
